+++ src/prefect/orion/exceptions.py
@@ -8,3 +8,7 @@ class ObjectNotFoundError(PrefectException):
     If thrown during a request, this exception will be caught and
     a 404 response will be returned.
     """
+
+
+class OrchestrationError(PrefectException):
+    """An error raised while orchestrating a state transition"""
